api/requirements.txt
groq
fastapi
uvicorn
tavily-python
